Added rack-cookie_rewrite.
Although the X-MSP-Header gets passed through to the various requests,
this is needed for the X-MSP-Set-Cookie headers.

We _could_ send back all headers in the same way we do the Set-Cookie
header, but I think the only one we actually care about is the
X-MSP-Set-Cookie, and this is already written.